Sauron Class. created math utilities file. migrated canvas functions to utils and sauron class.
