+++ public/js/project/canvas/canvas.js
@@ -11,9 +11,9 @@
   }
 * USAGE: var inputCanvas = Canvas(inputCanvasSettings);
 */
-// import vector.js
-var Vector = require('../actors/vector.js'),
-    Target = require('../actors/target.js');
+
+var Sauron = require('../sauron/sauron.js'),
+    utils = require('../utilities/math.js');
 
 function Canvas(settings) {
   //input error handling
@@ -40,10 +40,7 @@ Canvas.prototype.vectorDrag = function() {
                   x: d3.event.sourceEvent.x,
                   y: d3.event.sourceEvent.y
                 };
-                // in theory this would be Vector.updateInputVector(d);
-                updateInputVector(d);
-                updateOutputVector(d);
-                updateTargets(d);
+                Sauron.tellSauron(d);
               })
               .on("drag", function(d) {
               var d = {
@@ -52,9 +49,7 @@ Canvas.prototype.vectorDrag = function() {
                   y: d3.event.y
                 };
                 // in theory this would be Vector.updateInputVector(d);
-                updateInputVector(d);
-                updateOutputVector(d);
-                updateTargets(d);
+                Sauron.tellSauron(d);
               });
 
     return drag;
@@ -77,6 +72,10 @@ Canvas.prototype.getDefs = function(type) {
   return d3.select('#'+id+'-defs');
 };
 
+Canvas.prototype.getTar = function() {
+  return d3.select('#tar_img');
+};
+
 Canvas.prototype.appendSvg = function(type) {
   var id = type || this.type;
   var canvas = this.getCanvas(id).append('svg')
@@ -97,10 +96,6 @@ Canvas.prototype.addImage = function() {
          "height": "40",
          "xlink:href": "../public/img/target.gif"
        });
-};
-
-Canvas.prototype.getTar = function() {
-  return d3.select('#tar_img');
 };
 
 Canvas.prototype.appendImageToPattern = function() {
@@ -173,60 +168,6 @@ Canvas.prototype.drawProgressBar = function() {
          .text("0% Complete");
 }
 
-function updateInputVector(d){
-// redraw input vector
-  d3.select('#input-vector').remove();
-  d3.select('#input-svg').append('path')
-    .attr({
-      "stroke": "red",
-      "stroke-width":"4",
-      "d": "M 250 250 L"+d.x+" "+d.y+"z",
-      "id": 'input-vector'
-  });
-};
-
-function updateOutputVector(d) {
-  var i = applyMatrix(d.x,d.y);
-  d3.select('#output-vector').remove();
-  d3.select('#output-svg').append('path')
-    .attr({
-      "stroke": "red",
-      "stroke-width":"4",
-      //"fill": "/path/to/here",
-      "d": "M 250 250 L"+i[0]+" "+i[1]+"z",
-      "id": 'output-vector'
-  });
-};
-
-function updateTargets(d) {
-  var x = d3.selectAll("circle").attr("cx"),
-      y = d3.selectAll("circle").attr("cy"),
-      r = d3.selectAll("circle").attr("r"),
-      i = applyMatrix(d.x,d.y);
-  if (isClose(i[0],i[1],x,y,r)) {
-    d3.selectAll("circle").remove();
-    updateProgress();
-    generateTarget([[1,3],[2,0]]);
-  }
-}
-
-function updateProgress() {
-    var bar = d3.select('#progressbar'),
-        score = d3.select('#score');
-        curr = bar.attr("aria-valuenow");
-        if (Number(curr) >= 100) {
-          curr = 100;
-          score.text("Proceed To Next Level!");
-        }
-        else {
-          curr = Number(curr) + 5;
-          score.text(curr + "% Complete");
-        }
-
-        bar.style("width", curr + "%");
-        bar.attr("aria-valuenow", curr);
-}
-
 /**
 *
 *
@@ -241,7 +182,7 @@ Canvas.prototype.checkCollisions = function(oldVector,newVector,targets) {
     var tar_x = targets[i].x,
         tar_y = targets[i].y;
     //10 pixels is from Joseph's old settings
-    if (isClose(end_x,end_y,tar_x,tar_y,10)) {
+    if (utils.isClose(end_x,end_y,tar_x,tar_y,10)) {
       results[i] = true;
     }
     else {
@@ -251,7 +192,7 @@ Canvas.prototype.checkCollisions = function(oldVector,newVector,targets) {
       var dis_x = start_x + param * (end_x - start_x);
       var dis_y = start_y + param * (end_y - start_y);
       //10 pixels is from Joseph's old settings
-      if (isClose(dis_x,dis_y,tar_x,tar_y,10)) {
+      if (utils.isClose(dis_x,dis_y,tar_x,tar_y,10)) {
         results[i] = true;
       }
       else {
@@ -262,44 +203,35 @@ Canvas.prototype.checkCollisions = function(oldVector,newVector,targets) {
   return results;
 }
 
-
-function isClose(oX, oY, tX, tY, radius) {
-  var dist = Math.sqrt(Math.pow((tX - oX),2) + Math.pow((tY - oY),2));
-  if (dist <= radius) {
-    return true;
-  }
-  return false;
-}
-
 Canvas.prototype.proximity = function(outputVector, target) {
-    if (isClose(targetX, targetY, 500)) {
+    if (utils.isClose(targetX, targetY, 500)) {
       target.updateColor('#e5e5ff', target.id);
     }
-    else if (isClose(targetX, targetY, 450)) {
+    else if (utils.isClose(targetX, targetY, 450)) {
       target.updateColor('#ccccff', target.id);
     }
-    else if (isClose(targetX, targetY, 400)) {
+    else if (utils.isClose(targetX, targetY, 400)) {
       target.updateColor('#b2b2ff', target.id);
     }
-    else if (isClose(targetX, targetY, 350)) {
+    else if (utils.isClose(targetX, targetY, 350)) {
       target.updateColor('#9999ff', target.id);
     }
-    else if (isClose(targetX, targetY, 300)) {
+    else if (utils.isClose(targetX, targetY, 300)) {
       target.updateColor('#7f7fff', target.id);
     }
-    else if (isClose(targetX, targetY, 250)) {
+    else if (utils.isClose(targetX, targetY, 250)) {
       target.updateColor('#6666ff', target.id);
     }
-    else if (isClose(targetX, targetY, 200)) {
+    else if (utils.isClose(targetX, targetY, 200)) {
       target.updateColor('#4c4cff', target.id);
     }
-    else if (isClose(targetX, targetY, 150)) {
+    else if (utils.isClose(targetX, targetY, 150)) {
       target.updateColor('#3232ff', target.id);
     }
-    else if (isClose(targetX, targetY, 100)) {
+    else if (utils.isClose(targetX, targetY, 100)) {
       target.updateColor('#1919ff', target.id);
     }
-    else if (isClose(targetX, targetY, 50)) {
+    else if (utils.isClose(targetX, targetY, 50)) {
       target.updateColor('#0000ff', target.id);
     }
   }
@@ -319,62 +251,8 @@ Canvas.prototype.drawTarget = function(target) {
     });
 };
 
-Canvas.prototype.drawTargets = function(targets) {
-  // just a for loop with drawTarget
-  for( var i = 0; i < targets.length; i++ ) {
-    drawTarget(targets[i]);
-  }
-};
-
 Canvas.prototype.checkProximity = function(vector, target) {
-  return this.isClose(vector.head.x, vector.head.y, target.x, target.y, target.r);
-}
-
-function screenToMath(x,y) {
-  return [(x - 250) * 10 / 250, - (y - 250) * 10 / 250];
-}
-
-function mathToScreen(x,y) {
-  return [x * 250 / 10 + 250, - y * 250 / 10 + 250];
-}
-
-function applyMatrix(sX,sY,matrix) {
-  var matrix = matrix || [[1,3],[2,0]];
-  var math_coord = screenToMath(sX,sY),
-      applied_coord = [matrix[0][0] * math_coord[0] + matrix[0][1] * math_coord[1], matrix[1][0] * math_coord[0] + matrix[1][1] * math_coord[1]];
-  return mathToScreen(applied_coord[0],applied_coord[1]);
-}
-
-function getRandom(min,max) {
-  return Math.random() * (max - min) + min;
-}
-
-function generateTarget(matrix) {
-  var legal = false,
-      par = matrix[0][0] * matrix[1][1] - matrix[0][1] * matrix[1][0];
-  var newX, newY;
-
-  while (!legal) {
-    newX = getRandom(0,500);
-    newY = getRandom(0,500);
-    var pre = screenToMath(newX,newY);
-    var prex = (matrix[1][1] * pre[0] - matrix[0][1] * pre[1]) / par,
-        prey = (- matrix[1][0] * pre[0] + matrix[0][0] * pre[1]) / par;
-    pre = mathToScreen(prex,prey);
-
-    if (pre[0] >= 0 && pre[0] <= 500 && pre[1] >= 0 && pre[1] <= 500) {
-      legal = true;
-      var targetSettings = {
-      	x: newX,
-      	y: newY,
-      	r: 20,
-      	color: "black",
-      	isScore: false
-      };
-      var newTarget = new Target(targetSettings);
-      newTarget.drawTarget();
-    }
-  }
+  return utils.isClose(vector.head.x, vector.head.y, target.x, target.y, target.r);
 }
 
 module.exports = Canvas;
